+++ lib/legion/extensions/apollo.rb
@@ -24,3 +24,20 @@ module Legion
     end
   end
 end
+
+# Entity watchdog on post_tick_reflection
+if defined?(Legion::Gaia::PhaseWiring) && begin
+  Legion::Settings.dig(:apollo, :entity_watchdog, :enabled)
+rescue StandardError
+  false
+end
+  require 'legion/extensions/apollo/helpers/entity_watchdog'
+  Legion::Gaia::PhaseWiring.register_handler(:post_tick_reflection) do |tick_results|
+    text = tick_results.is_a?(Hash) ? (tick_results[:content] || tick_results[:output] || '').to_s : tick_results.to_s
+    entities = Legion::Extensions::Apollo::Helpers::EntityWatchdog.detect_entities(text: text)
+    if entities.any?
+      Legion::Extensions::Apollo::Helpers::EntityWatchdog.link_or_create(entities:       entities,
+                                                                         source_context: tick_results[:tick_id])
+    end
+  end
+end